Added input data validation through INotifyDataErrorInfo

+++ CompareSearchMethods/Model/ISimulationResults.cs
@@ -4,7 +4,7 @@ namespace CompareSearchMethods.Model
 	public interface ISimulationResults
 	{
 		int NoOfEntries { get; set; }
-		long NoOfSearches { get; set; }
+		int NoOfSearches { get; set; }
 		double AvgNoOfIterations { get; set; }
 		double AvgElapsedTime { get; set; }
 	}
